added crud for managers

+++ app/Http/Controllers/BooksController.php
@@ -16,9 +16,9 @@ class BooksController extends Controller
      */
     public function index()
     {
-        $books= books::all();
+        $books = books::all();
 
-        return view('Books.list',compact( "books"));
+        return view('Books.list', compact("books"));
     }
 
     /**
@@ -36,24 +36,26 @@ class BooksController extends Controller
     {
         // $book->title=$request->title;
         // $book->description=$request->description;
-        $book=new books($request->all());
-        if($request->hasFile('Image')){ 
-            $destination_path ='public/images/books';
-            $image=$request->file('Image');
-            $image_name=$image->getClientOriginalName();
-            $path=$request->file('Image')->storeAs($destination_path,$image_name);
-            $book['Image']=$image_name;
+        $book = new books($request->all());
+
+        $this->authorize('create', $book);
+        if ($request->hasFile('Image')) {
+            $destination_path = 'public/images/books';
+            $image = $request->file('Image');
+            $image_name = $image->getClientOriginalName();
+            $path = $request->file('Image')->storeAs($destination_path, $image_name);
+            $book['Image'] = $image_name;
         }
         // $this->authorize('create', books::class);
-    //     $validaated=$request->validate(['name'=>'required', 
-    //     'title'=>'required |numeric',
-    //     'description'=>'required |unique:students'
-    //     ,'tracks_id'=>'required']
-    // );        
+        //     $validaated=$request->validate(['name'=>'required', 
+        //     'title'=>'required |numeric',
+        //     'description'=>'required |unique:students'
+        //     ,'tracks_id'=>'required']
+        // );        
         // $st->user_id = auth()->user()->id;
 
         $book->save();
-        return to_route('books.index')-> with('success',' create and save success');
+        return to_route('books.index')->with('success', ' create and save success');
     }
 
     /**
@@ -69,16 +71,16 @@ class BooksController extends Controller
      */
     public function edit(books $book, Request $request)
     {
-    // dd($book);
-    //  $this->authorize('update', $book);
+        // dd($book);
+        //  $this->authorize('update', $book);
 
-     // if ($request->user()->cannot('update', $Student)) 
-     // return "no access";
- 
-  
-     return view('Books.edit',compact("book"));// لازم نفس اسم المتغير وبرده بنادي عليه بنفس الاسم في الليست
+        // if ($request->user()->cannot('update', $Student)) 
+        // return "no access";
 
-     //
+
+        return view('Books.edit', compact("book")); // لازم نفس اسم المتغير وبرده بنادي عليه بنفس الاسم في الليست
+
+        //
         //
     }
 
@@ -88,28 +90,31 @@ class BooksController extends Controller
     // public function update(StorebooksRequest $request, books $book)
     public function update(UpdatebooksRequest $request, books $book)
     {
-    $book->update($request->all());
-    if($request->hasFile('Image')){ 
-        $destination_path ='public/images/books';
-        $image=$request->file('Image');
-        $image_name=$image->getClientOriginalName();
-        $path=$request->file('Image')->storeAs($destination_path,$image_name);
-        $book['Image']=$image_name;
-        }
-    $book->save();
-    return to_route('books.index');
+        $book->update($request->all());
 
+        $this->authorize('update', $book);
+        if ($request->hasFile('Image')) {
+            $destination_path = 'public/images/books';
+            $image = $request->file('Image');
+            $image_name = $image->getClientOriginalName();
+            $path = $request->file('Image')->storeAs($destination_path, $image_name);
+            $book['Image'] = $image_name;
+        }
+        $book->save();
+        return to_route('books.index');
     }
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(  Request $request ,books $book)
+    public function destroy(Request $request, books $book)
     {
-    //  $books= books::find($books);
-     $book->delete();
-     $request->session()->flash('success',"deleted successfully");
 
-     return to_route('books.index');
+        $this->authorize('delete', $book);
+        //  $books= books::find($books);
+        $book->delete();
+        $request->session()->flash('success', "deleted successfully");
+
+        return to_route('books.index');
 
         //
     }
